+++ packages/forma-36-website/gatsby-node.js
@@ -1,8 +1,6 @@
 const fs = require('fs');
 
 exports.onPostBuild = () => {
-  fs.copyFileSync('./_redirects', './public/_redirects', err => {
-    if (err) throw err;
-    console.log('redirect added!'); // eslint-disable-line
-  });
+  fs.copyFileSync('./_redirects', './public/_redirects');
+  console.log('redirect added!'); // eslint-disable-line
 };