+++ packages/forma-36-website/src/components/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TextLink } from '@contentful/forma-36-react-components';
-import { Flex } from '@contentful/forma-36-react-components/dist/alpha';
+import { Flex } from '@contentful/forma-36-react-components';
 import tokens from '@contentful/forma-36-tokens';
 import { css } from '@emotion/core';