+++ packages/forma-36-website/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.2](https://github.com/contentful/forma-36/compare/forma-36-website@0.11.1...forma-36-website@0.11.2) (2021-01-05)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.11.1](https://github.com/contentful/forma-36/compare/forma-36-website@0.11.0...forma-36-website@0.11.1) (2021-01-05)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+# [0.11.0](https://github.com/contentful/forma-36/compare/forma-36-website@0.10.12...forma-36-website@0.11.0) (2021-01-04)
+
+
+### Features
+
+* export previous alpha components as stable ([ba1f919](https://github.com/contentful/forma-36/commit/ba1f91949be0517a2a03567bc8b3588037f5d021))
+
+
+
+
+
+## [0.10.12](https://github.com/contentful/forma-36/compare/forma-36-website@0.10.11...forma-36-website@0.10.12) (2021-01-04)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
 ## [0.10.11](https://github.com/contentful/forma-36/compare/forma-36-website@0.10.10...forma-36-website@0.10.11) (2021-01-04)
 
 **Note:** Version bump only for package forma-36-website